* restore-src/restore.c: Fix read_to_connection argument.
git-svn-id: https://amanda.svn.sourceforge.net/svnroot/amanda/amanda/trunk@2319 a8d146d6-cc15-0410-8900-af154a0219e0

+++ restore-src/restore.c
@@ -1298,7 +1298,7 @@ void restore(RestoreSource  *source,
 
     if (myout->restore_mode == LOCAL_DIRECTTCP_MODE ||
 	myout->restore_mode == DIRECTTCP_MODE) {
-	gsize size = 0;
+	guint64 size = 0;
 
 	if (!is_continuation) {
 	    device_accept(source->u.device,
